Moved code from routes to status-controller
